Encore integers using signed leb128 - fixes #75.

+++ src/wasm/sections.rs
@@ -33,9 +33,9 @@ impl SectionType {
             }
 
             fun_type.push(FUNC);
-            fun_type.append(&mut to_leb(params.len()));
+            fun_type.append(&mut to_leb(params.len() as u64));
             fun_type.extend(params);
-            fun_type.append(&mut to_leb(results.len()));
+            fun_type.append(&mut to_leb(results.len() as u64));
             fun_type.extend(results);
 
             match known_types.get(&fun_type) {
@@ -53,12 +53,12 @@ impl SectionType {
             }
         }
 
-        let count = to_leb(types.len());
+        let count = to_leb(types.len() as u64);
         size += count.len();
 
         SectionType {
             types: types,
-            size: to_leb(size),
+            size: to_leb(size as u64),
             count: count,
         }
     }
@@ -93,17 +93,17 @@ impl SectionFunction {
 
         // The function index corresponds to its index in funs
         for fun in funs {
-            let idx = to_leb(fun.type_idx);
+            let idx = to_leb(fun.type_idx as u64);
             size += idx.len();
             types.push(idx);
         }
 
-        let count = to_leb(types.len());
+        let count = to_leb(types.len() as u64);
         size += count.len();
 
         SectionFunction {
             types: types,
-            size: to_leb(size),
+            size: to_leb(size as u64),
             count: count,
         }
     }
@@ -134,23 +134,23 @@ struct SectionMemory {
 impl SectionMemory {
     fn new(memories: Vec<wasm::Limit>) -> SectionMemory {
         let mut mem = Vec::new();
-        let count = to_leb(memories.len());
+        let count = to_leb(memories.len() as u64);
 
         for memory in memories {
             match memory {
                 wasm::Limit::Min(min) => {
                     mem.push(0x00); // No upper limit flag
-                    mem.extend(to_leb(min as usize));
+                    mem.extend(to_leb(min as u64));
                 }
                 wasm::Limit::MinMax(min, max) => {
                     mem.push(0x01); // With upper limit flag
-                    mem.extend(to_leb(min as usize));
-                    mem.extend(to_leb(max as usize));
+                    mem.extend(to_leb(min as u64));
+                    mem.extend(to_leb(max as u64));
                 }
             }
         }
 
-        let size = to_leb(mem.len() + count.len());
+        let size = to_leb((mem.len() + count.len()) as u64);
         SectionMemory {
             memories: mem,
             count: count,
@@ -186,19 +186,19 @@ impl SectionCode {
         for fun in funs {
             let body = &fun.body;
 
-            let mut size_body = to_leb(body.len());
+            let mut size_body = to_leb(body.len() as u64);
             size_body.extend(body);
 
             size += size_body.len();
             fun_bodies.push(size_body);
         }
 
-        let count = to_leb(fun_bodies.len());
+        let count = to_leb(fun_bodies.len() as u64);
         size += count.len();
 
         SectionCode {
             count: count,
-            size: to_leb(size),
+            size: to_leb(size as u64),
             bodies: fun_bodies,
         }
     }
@@ -236,22 +236,22 @@ impl SectionExport {
                 let mut data = Vec::new();
                 let encoded_name = name.as_bytes();
 
-                data.extend(to_leb(encoded_name.len()));
+                data.extend(to_leb(encoded_name.len() as u64));
                 data.extend(encoded_name);
                 data.push(KIND_FUNC);
-                data.extend(to_leb(idx));
+                data.extend(to_leb(idx as u64));
 
                 size += data.len();
                 exports.push(data);
             }
         }
 
-        let count = to_leb(exports.len());
+        let count = to_leb(exports.len() as u64);
         size += count.len();
 
         SectionExport {
             exports: exports,
-            size: to_leb(size),
+            size: to_leb(size as u64),
             count: count,
         }
     }
